staging: dwc2: Fix naming of register constants
A generic set of FIFOSIZE_* constants was used for both the GNPTXFSIZ
and HPTXFSIZ registers, even where a specific set of constants was a
available for the GNPTXFSIZ register. This change adds specific
constants for the HPTXFSIZ register as well, changes the code to use
those specific constants and removes the generic constants.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>