staging: dwc2: Set the utmi_width param to 16 by default
The fixed value in pci.c is 16 and the comments about dwc2_core_params
also says the default should be 16. However, the default in
dwc2_set_param_phy_utmi_width, used when the dwc2_core_params contains
an invalid value, is now 8.

This commits changes that default to 16, as documented. Given there is
currently only the pci binding and there is no way to change the value
there, except for changing pci.c and recompiling, this change shouldn't
actually change any behaviour with the current driver.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>